// src/app/interface/interface.ts
export interface LostItem{
    _id:string,
      title: string;
      image: string;
      location: boolean;
      description: boolean;
      identifying_question: string;
      claimed: boolean;
      contact_name: string;
      email: string;
      phonenumber: string;
      date:string;
      date_found: string;



      // 今後修正するinterface
      // Title:string;
      // Image: string;
      // Location:string;
      // Description:string;
      // Date_found:string;
      // Identifying_question: string;
      // Claimed:boolean;
      // Contact_name: string;
      // Email : string;
      // Phonenumber:string;

}